+++ __tests__/bundle/bundle-lint-format/codeframe-format-snapshot.js
@@ -2,8 +2,6 @@
 
 exports[`E2E bundle lint format bundle lint: should be formatted by format: codeframe 1`] = `
 
-No configurations were defined in extends -- using built in recommended configuration by default.
-
 [1] openapi.yaml:20:11 at #/paths/~1my_post/post/requestBody/content/application~1json
 
 Expected type \`MediaType\` (object) but got \`null\`