+++ packages/cli/src/__tests__/commands/push.test.ts
@@ -7,7 +7,12 @@ import {
 } from '../../commands/push';
 
 jest.mock('fs');
-jest.mock('node-fetch');
+jest.mock('node-fetch', () => ({
+  default: jest.fn(() => ({
+    ok: true,
+    json: jest.fn().mockResolvedValue({}),
+  })),
+}));
 jest.mock('@redocly/openapi-core');
 jest.mock('../../utils');
 
@@ -16,7 +21,7 @@ jest.mock('../../utils');
 describe('push', () => {
   const redoclyClient = require('@redocly/openapi-core').__redoclyClient;
 
-  beforeAll(() => {
+  beforeEach(() => {
     jest.spyOn(process.stdout, 'write').mockImplementation(() => true);
   });