+++ packages/core/src/rules/common/__tests__/operation-4xx-response.test.ts
@@ -21,7 +21,7 @@ describe('Oas3 operation-4xx-response', () => {
     const results = await lintDocument({
       externalRefResolver: new BaseResolver(),
       document,
-      config: makeConfig({ 'operation-4xx-response': 'error' }),
+      config: await makeConfig({ 'operation-4xx-response': 'error' }),
     });
 
     expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`
@@ -60,7 +60,7 @@ describe('Oas3 operation-4xx-response', () => {
     const results = await lintDocument({
       externalRefResolver: new BaseResolver(),
       document,
-      config: makeConfig({ 'operation-4xx-response': 'error' }),
+      config: await makeConfig({ 'operation-4xx-response': 'error' }),
     });
 
     expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`Array []`);
@@ -83,7 +83,7 @@ describe('Oas3 operation-4xx-response', () => {
     const results = await lintDocument({
       externalRefResolver: new BaseResolver(),
       document,
-      config: makeConfig({ 'operation-4xx-response': 'error' }),
+      config: await makeConfig({ 'operation-4xx-response': 'error' }),
     });
 
     expect(replaceSourceWithRef(results)).toMatchInlineSnapshot(`